Do not fail when repoquery is not installed
A good chunk of container images try to be as slim as possible and as
a consequence do not have repoquery binary which is used to limit the
packages to a given repo.

This commit makes the playbook more against these containers and
installs all available updates when repoquery is not installed, the
same as when we don't restrict the list of repositories to get updates
from.

Change-Id: I81e5cd0edae56c1535a2e4dcc639ec71055b10d2