Add rpm_install tasks
This patch adds a new rpm_install tasks which can be used
to build a new container layer from a set of local RPMs.
This is useful for building contains locally with no network
connectivity.

As part of the change the existing Yum update Dockerfile.j2 was
renamed to Dockerfile-yum.j2 to make it more specific.

Change-Id: I951c976ebb84c28cecefd2f5753a8232f46ad4aa