Add Buildah support
Buildag is a tool that facilitates building OCI container images.
This patch adds support to modify and update containers built by
Buildah.
It's just making sure we can run the build command with their CLI and
also get the user from the container config.
The default remains 'docker' for backward compatibility.

Note: it push the new images with sudo to avoid permissions errors.

Change-Id: I3db1934ad826ec6433bd6aec067112e91d82e355