Only do yum update when needed
Currently the logic for this results in a full yum update when no
package updates are found in the provided repository. This can lead to
job timeouts when nothing was built in CI because it effectively does a
yum update on every container and applies other system packages rather
than ones that actually changed.

There is a larger issue in that we can still get out of sync with the
host OS.

Change-Id: Iaf41691ea3cb6e78186741ac5e15614fb73f89ff
Related-Bug: #1786764