Don't yum update files which are not installed
This change builds the packages_for_update list by intersecting
installed packages with the packages available, which avoids passing
uninstalled packages to the yum update call.

For some reason doing a yum update with packages which are not
installed causes a delay for each uninstalled package.

For my test case this gave a speed improvement of 35s vs 45s for the
whole image build. This would accumulate to decent time savings for
all images in CI.

Change-Id: I338d9881d5a05dd7b1fbc5efea1145349f0997a6