Fix local (non-root) docker build support
This was broken recently in a2229e738ae1d52cfa39b00972925e505fdca292
where the buildah support hard coded the modify command so that
it required a 'sudo' for the docker version.

Docker does not always require root for these commands. Also,
need to look further as to why buildah required this.

Change-Id: Iae231ec8bc8dc2822fc3d6bfab593cff3c599910