Use a tempfile for the modified Dockerfile
When using modify_image.yml directly with a multiple image prepare,
the Dockerfile.modified can be overwritten multiple times in a race
which causes the wrong images to be built.

This change uses a tempfile for the modified Dockerfile, avoiding this
issue.

yum_update.yml and rpm_install.yml is not affected by this because
they write out the Dockerfile to a temp directory for each image.

Change-Id: I374f76650ef0b0211d92435d66fdead2b7730b21