Cleanup buildah working container
Currently in the yum update output, we see buildah trying to rmi the
image we were working with however it is currently in use. After we
commit our changes we need to cleanup our working container so we can
remove the image we were using (if not used by another process
elsewhere).

Change-Id: I54e37b43346b97be0a7cfab12e6cac9809537c83