Enable saving of rpms
We currently map in the cache dir, but it only saves the metadata. The
improvements would come with the rpm saving as well. In order to have
that, we need to set keepcache=1 which is disabled by default. We can
enable this on the command line via --setopt so we don't have to touch
the container yum configuration.

Change-Id: Ia274dbbad9a220e52b2cfa554dede3bf40055f98