Use length to avoid unsupported operations with '>'
The dev_install.yaml script breaks when running python 3 because of
unsupported comparisions between 'AnsibleUnsafeText' and 'int' with the
'>' operator.

This is very similar to a bug opened against ansible upstream [0]. We
can fix the issue by using length explicitly, which ensures we're
comparing integers to integers since we're dealing with a list.

[0] https://github.com/ansible/ansible/issues/50388

Change-Id: I100639982b75c9d345269ba9bf10defb7e96d774