Ensure the yum cache has at most one writer
If the yum cache path exists and already mounted by someone,
do not attempt writing to it, use the overlay mode instead.

This still leaves a window of opportunity for another workers to
RW mount the cache after the ansible check has reported a stale
fact about there was no other mounts found. But this is unlikely
to happen.

Also, if it has to be retried in the rescue block, do not use the yum
cache for the maximum data safety and clean (a scratch) cache state
reasons.

This drastically reduces chances to have multiple writers for the
cache.

Closes-bug: #1860804

Change-Id: I19491a162e5bf6d6517fd343d675aff12bdc9719
Signed-off-by: Bogdan Dobrelya <bdobreli@redhat.com>