Add some notes about SELinux and limitations
With newer podman and container-selinux, we may face some issues
depending on where directories are located. The way we're running the
role is also important - being launched as root or as a user makes a big
difference, especially with the user's home content.

It also updates the setup.cfg content to match new format requirements

Change-Id: Ib2b7fd2f557d4743efd2eaca18474fb45b91cbcf